Min fix: return type in getMarginTop from int to float.

+++ src/PaymentPart/Output/Element/Placeholder.php
@@ -113,7 +113,7 @@ class Placeholder implements OutputElementInterface
         return $this->float;
     }
 
-    public function getMarginTop(): int
+    public function getMarginTop(): float
     {
         return $this->marginTop ?: 0;
     }
